Changes pointer to SdFile to SdFile member object
The SD library was leaking memory due to the lack of destructor. There
were subtle situations where temporary objects were created and
destructed and the programmer had no chance to call close(). So, we
need a destructor that calls close(). But to do so, we also need to
change the pointer to SdFile to an actual SdFile object, otherwise the
pointer will be copied and when one object is freed, the other object
will loose its buffer too.