Use correct include syntax
Angle brackets for external includes. Quotes for local includes. The former is not critical but the latter is necessary in order to use the library when it's not in one of the standard Arduino libraries folders (such as when bundled with a sketch), where local files included with the angle brackets syntax will not be found in the include search path, causing compilation to fail.