Fix position of c_size_high in struct csd2_t
Byte 7 of register CSD v2.00 is so composed: the first 2 bits for "reserved3" and the remaining 6 bits for "c_size_high". Because of how the bits are arranged on little-endian systems, "c_size_high" must be positioned before "reserved3" in order to have "c_size_high", "c_size_mid" and "c_size_low" arraged contiguously.