Format demo code in reference
Some of the library reference pages provide code demonstrating usage. In some places, this code had a very odd
formatting where the first indent was three spaces, followed by the standard two space indent for the rest. This made
the structure of the code difficult to follow and also results in an unexpected diff when the user runs the formatter on the
code in Arduino IDE or Web Editor. It also makes it unfriendly to contributors.

All code in the reference is hereby formatted to be compliant with the Arduino IDE 2.x auto formatter (which uses a
configuration equivalent that of the formatters in Arduino IDE 1.x and Arduino Web Editor).