Avoid unnecessary calls in `rewindDirectory` reference code
The reference page for `File::rewindDirectory` provides a sketch demonstrating its usage.

This code calls the function after iterating over each directory on the card. This would be necessary if the
`printDirectory` function was configured to iterate over that directory multiple times, but it is not. So the calls are
redundant.

Moving the call to the `setup` function has the following benefits:

- Makes the use of the function more prominent. The reader can get an overview of what the program is doing without
  delving into the `printDirectory` function
- Avoids the misconception that it is required to call the function recursively as was done before