Print the card file structure twice in `rewindDirectory` reference code
`File::rewindDirectory` is used to reset the current position in the file system. Previously, the file structure was
only iterated over once, in which case rewinding is not necessary and will have no observable effect.

In the following sequence, it is necessary and does have an observable effect:

1. Iterate over file structure using `File::openNextFile`
2. Reset current position using `File::rewindDirectory`
3. Iterate over file structure using `File::openNextFile`

The reader can simply remove the `File::rewindDirectory` call from the program that follows this sequence for a clear
demonstration of its purpose.