Fix spell check false positive
The "codespell" tool is used to check for commonly misspelled words in the project files.

As part of a recent update, "indx" was added to the misspelled words dictionary as a misspelling of "index". The project
contains a variable named "indx", so the spell check started failing following that update. Although it is questionable
whether using a variable name that is shortened from the related correct word by only a single letter, it is also not
objectively wrong. Since it is believed that this was an intentional choice by the developer, the spell check failure
can be considered spurious. In this case, the correct resolution is to configure codespell to ignore this word.